text cleaning
added text cleaning at /clean_text endpoint
it takes in a query, and replies with a json of similar format with the lowercase lemmas without stopwords.
