+++ Project 2/app/routes/__init__.py
@@ -1,3 +1,3 @@
-from .llm_classifier import router as classify_router
+from .classifier import router as classify_router
 
 __all__ = ["classify_router"]
