+++ Project 2/requirements.txt
@@ -6,3 +6,5 @@ fastapi[standard]>=0.113.0, <0.114.0
 pydantic>=2.7.0, <3.0.0
 scikit-learn>=1.6.1
 httpx>=0.28.1
+spacy>=3.0.0,<4.0.0
+en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
